+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Symbol rotation [#294](https://github.com/mapsforge/vtm/issues/294)
 - Marker clustering [#312](https://github.com/mapsforge/vtm/issues/312)
 - Osmagray theme [#300](https://github.com/mapsforge/vtm/issues/300)
+- Location renderer custom shaders [#317](https://github.com/mapsforge/vtm/issues/317)
 - OkHttp external cache [#135](https://github.com/mapsforge/vtm/issues/135)
 - Texture atlas improvements [#301](https://github.com/mapsforge/vtm/pull/301) [#304](https://github.com/mapsforge/vtm/pull/304)
 - Many other minor improvements and bug fixes